requirements.txt
PyAutoGUI
PyYAML
pillow
pytesseract
